Changes to add custom equality checking
Added method overloads to allow the supply of two functions to check the
equality of an item.

keyEqualityFunction - To find the new item within the original list.
equalityFunction - To determine whether or not the item has changed.

In some cases an object might not inplement full equality and only want
to validate whether an item exists in the original list by checking a
specific piece of information e.g. Key information. Also the editable
information may not be included in the Object.Equals equality check.